requirements.txt
python-dotenv==0.15.0
requests==2.24.0
scispacy==0.2.5
spacy==2.3.0
spacy-streamlit==0.0.2
streamlit==0.62.1
